requirements.txt
# Web Framework
flask>=2.0.3
flask-cors>=4.0.0

# Environment & Configuration
python-dotenv>=0.19.2

# Authentication
authlib>=1.0

# HTTP Requests
requests>=2.27.1

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.17.0

# API Documentation
flask-restx>=1.1.0
flasgger>=0.9.7